### 2024-02-19 #81 @celinesoeiro
[Re] Cortical network effects of subthalamic deep brain stimulation in a thalamo-cortical microcircuit model
**Original article:** AmirAli Farokhniaee and Madeleine M Lowery 2021 'Cortical network effects of subthalamic deep brain stimulation in a thalamo-cortical microcircuit model', Journal of Neural Engineering, DOI: 10.1088/1741-2552/abee50

**Scientific domain:** Computational Neuroscience
