### 2024-04-28 eb924c9
update readme + general requested fixes
