+++ requirements-conda.txt
@@ -1,0 +1,437 @@
+# This file may be used to create an environment using:
+# $ conda create --name <env> --file <this file>
+# platform: osx-arm64
+@EXPLICIT
+https://repo.anaconda.com/pkgs/main/osx-arm64/fftw-3.3.9-h1a28f6b_1.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/blas-1.0-openblas.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/bzip2-1.0.8-h620ffc9_4.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/c-ares-1.18.1-h1a28f6b_0.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/ca-certificates-2023.01.10-hca03da5_0.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/giflib-5.2.1-h80987f9_3.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/jpeg-9e-h80987f9_1.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/jxrlib-1.1-h1a28f6b_2.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/ldid-2.1.2-h64d1936_2.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/libbrotlicommon-1.0.9-h1a28f6b_7.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/libcxx-14.0.6-h848a8c0_0.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/libdeflate-1.17-h80987f9_0.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/libev-4.33-h1a28f6b_1.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/libffi-3.4.2-hca03da5_6.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/libiconv-1.16-h1a28f6b_2.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/libsodium-1.0.18-h1a28f6b_0.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/libuv-1.44.2-h80987f9_0.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/libwebp-base-1.2.4-h80987f9_1.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/llvm-openmp-14.0.6-hc6e5704_0.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/lzo-2.10-h1a28f6b_2.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/ncurses-6.4-h313beb8_0.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/oniguruma-6.9.7.1-h1a28f6b_0.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/patch-2.7.6-h1a28f6b_1001.conda
+https://repo.anaconda.com/pkgs/main/noarch/tzdata-2022g-h04d1e81_0.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/xz-5.2.10-h80987f9_1.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/yaml-0.2.5-h1a28f6b_0.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/zlib-1.2.13-h5a0b063_0.conda
[…]
