+++ tcm_params.py
@@ -101,12 +101,12 @@ def TCM_model_parameters():
     dbs_end = int(dbs_begin*2)              # position where DBS stops being applied
         
     # Neuron quantities
-    qnt_neurons_s = 10         # Excitatory
-    qnt_neurons_m = 10         # Excitatory
-    qnt_neurons_d = 10         # Excitatory
-    qnt_neurons_ci = 10        # Inhibitory
-    qnt_neurons_tc = 10        # Excitatory
-    qnt_neurons_tr = 4         # Inhibitory
+    qnt_neurons_s = 100         # Excitatory
+    qnt_neurons_m = 100         # Excitatory
+    qnt_neurons_d = 100         # Excitatory
+    qnt_neurons_ci = 100        # Inhibitory
+    qnt_neurons_tc = 100        # Excitatory
+    qnt_neurons_tr = 40         # Inhibitory
     
     neuron_quantities = {
         'S': qnt_neurons_s,                      # Number of neurons in Superficial layer
