+++ dbs_compare.py
@@ -321,24 +321,27 @@ def dbs(freq):
     
     print("-- Done!")
     
-
+f_40, s_40 = dbs(25)
 f_80, s_80 = dbs(80)
 f_130, s_130 = dbs(130)
 f_180, s_180 = dbs(180)
 
-x_arr = np.arange(0, 81, 5)
+x_arr = np.arange(10, 41, 5)
 
 plt.figure(figsize=(21, 10))
-# plt.semilogy(f_PD, S_PD, color="steelblue")
+# plt.semilogy(f_DBS_PD, S_DBS_PD, color="steelblue")
 # plt.semilogy(f_normal, S_normal, color="darkorange")
+plt.semilogy(f_40, s_40, color="darkorchid")
 plt.semilogy(f_80, s_80, color="green")
 plt.semilogy(f_130, s_130, color="red")
 plt.semilogy(f_180, s_180, color="hotpink")
-plt.legend([ 'DBS - 80Hz', 'DBS - 130Hz', 'DBS - 180Hz'], fontsize=22)
-plt.ylim([1e-3, 1e8])
-plt.xlim([0, 80])
+# plt.legend([ 'PD', 'normal', 'DBS - 25Hz', 'DBS - 80Hz', 'DBS - 130Hz', 'DBS - 180Hz'], fontsize=22)
+plt.legend([ 'DBS - 25Hz', 'DBS - 80Hz', 'DBS - 130Hz', 'DBS - 180Hz'], fontsize=22)
+plt.ylim([1e-0, 1e6])
+plt.xlim([10, 40])
 plt.xticks(x_arr)
-plt.xlabel('frequency (Hz)')
-plt.ylabel('PSD [V**2/Hz]')
+plt.xlabel('frequency (Hz)', fontsize=22)
+plt.ylabel('PSD [V**2/Hz]', fontsize=22)
 plt.title('PSD', fontsize=22)
+plt.savefig('results/compare-PSD.png')
 plt.show()
