### 2024-05-21 #81 @celinesoeiro
@NikVard, we thank the reviewer for the time and effort devoted to our manuscript. We found his comments and suggestions helpful in improving the paper. We revised the manuscript, answering all points raised by the reviewer. Below, we reproduce the reviewer’s comments in boldface and our replies are in regular font.  The PDF file was also updated and can be found in the [in this link](https://github.com/celinesoeiro/TCM-model/blob/main/Neural_network_model_consisting_of_cortex_and_thalamus_to_simulate_effects_of_deep_brain_stimulation_in_Parkinson_s_disease.pdf).

**In Section 2.4 "Stimulation Protocol", you describe the "DBS off" vs the "DBS on" conditions in a simulation. In the original work, stimulation was applied for a total time T. Here, you stimulate for a time of T/3, which is 5 seconds for a simulation time of T=15 s. In the provided code of the original paper, they used a simulation time of 10 (+1) seconds to reach a steady state. Was it consistent?**
- In the original code (https://github.com/ModelDBRepository/266941/blob/main/ThalamoCortical_Microcircuit_PD_DBS_ModelDB.m), in line 190, one can see that 		when DBS is applied the authors divide the total time T in 3 equal parts. 			Regarding the +1 second to reach a steady state, we applied it at first but the 		results with or without it were very similar. Therefore we believe it is safe to say 		the results are consistent.

**It would have also been interesting to replicate the original authors' reported increase in beta-band power in response to low-frequency DBS (i.e. <40 Hz) and examine the results.**
- Done, with a frequency of 25 hz.

**The model currently plots figures as it runs, which are opened and block execution but never saved. Instead, opt for saving the figures in a dedicated "results" directory (i.e. using the code fig.savefig("out.png”)).**
- Done

**Updates in the PSD function to smoothen out your PSD estimates, you should set the parameter noverlap to get overlapping windows. Remove the nperseg.**
- Done

**Include installation instructions in the README Provide a conda .yml file, if conda was used, with the complete environment (including the Python version). Provide a requirements.txt file with the necessary modules to install using pip.**
- Done

**Update the parameters in tcm_params.py to match the ones in the paper**
- Done

**I can follow the code, but when I run the DBS.py file, there is an error in line 322 in the function plot_BP_filter**
- Thank you for pointing that out. We believe now you can run the code without this error.

**In Section 2.3 "Thalamo-cortical microcircuit model", you mention that "The values of synaptic strengths were adjusted to produce a simulated LFP in cortical layer D that resembles the LFP recorded under normal conditions". Could you specify how the original values were adjusted? How "close" was the simulated LFP to the recorded LFP and how was the similarity estimated?**
- In the original paper, the authors used Bayesian model comparison (BCM) to adjust the strengths for both normal and Parkinsonian conditions. For the Parkinsonian condition, they further adjusted the inhibitory strength from CI to D to increase the level of synchronous activity in layer D neurons. We added this explanation to the manuscript together with the cited references in the original article. In our reimplementation, we used the synaptic strengths given in the original code.

**In the first paragraph of page 4, you mention that "The scale of this threshold noise was set to one-third of the scale of ξ(t)". What does it mean "the scale of this threshold noise was set to one-third of ξ(t)"? Was it arbitrarily selected? In the original model, the mean of the threshold noise was set to half of the additive white noise.**
- This is due to a discrepancy between the original code and the original article. While in the original article the authors write “half of the additive white noise”, in the code they use one-third of the additive white noise (see lines 147 and 148 of 		the original code in ModelDB). We used the information given in the code because this provided a better fit to the plots in the original article. We added a comment about this discrepancy to the manuscript.

**In the last paragraph of page 3, you mention that the noise term was "scaled so that the mean firing rates of the neurons are compatible with experimental recordings;". Which recordings are you referring to? You can add the experimental reference from the original work here (I believe it's this one: "Li Q, Ke Y, Chan D C W, Qian Z M, Yung K K L, Ko H, Arbuthnott G and Yung W H 2012 Therapeutic deep brain stimulation in parkinsonian rats directly influences motor cortex Neuron 76 1030–41”.**
- Done

**The paragraph just below Table 2 is quite unclear. I think it would benefit greatly from rephrasing, especially the part starting with "for the excitatory neurons..."**
- We agree it was quite confuse. We have rephrased the paragraph, and hope it is clearer now.

**Table 1: you mention that the model contains 540 neurons in total, however, from the table, we see there are 600 neurons. From the code, I believe your TR should have 40 inhibitory neurons.**
- Thanks for pointing out this mistake to us. We have corrected the number of TR neurons in Table 1.

**First paragraph of "Methods": "Neurons of the S, M, S, and TC..." -> "Neurons of the S, M, D, and TC..."**
- Done
