### 2024-05-27 #81 @NikVard
@celinesoeiro Thank you for addressing the earlier comments and for the thorough description of the steps you took in response. After your edits, I can now run the code without issues and I will make some more comments that could be addressed.

# Code

1. First, thank you for addressing my comment and providing a requirements file and instructions to install the necessary python modules. I would like to point out that the conda environment you provided is OSX specific, and conda failed to find the relevant packages on other systems (linux and windows). You can follow the instructions from [this link](https://stackoverflow.com/questions/39280638/how-to-share-conda-environments-across-platforms) to make a non-platform-specific .yaml file that everyone can run if you wish.

2. As this is a computational work, could you consider specifying explicitly in the text the simulation parameters that you used to produce the figures? I believe there may be some errors in the code that need to be addressed, which I will detail in the following points. It is also good practice to include the parameters  in the text to encourage the reproducibility of your work by others who read the paper alone. Could you kindly specify the simulation parameters in a table?

3. I noticed something strange in the declaration of your parameters. In the `tcm_params.py` file, the default time step is initialized to `100/ms`; that is 100ms, not 10ms as stated in the comments. Can you clarify if you ran all simulations with a time step of 10ms or 100ms?

4. Unless I am mistaken, there are also some follow-up issues with the initialization of other variables. Following point 3 from above, in the `tcm_params.py` file you define dt as 100/ms, which, as stated above, evaluates to 0.1 _seconds_. Then, you proceed to define your sampling frequency as `samp_freq = int(ms/dt) # sampling frequency in Hz`, which gives a value of 10000. This corresponds to a **10kHz** sampling frequency (and a dt of 0.1 _ms_). This is inconsistent with what you probably intended to do, as with a dt of 0.1 seconds (100ms equivalently) you would only have a sampling frequency of 10Hz. This is evident also by evaluating the expression you wrote in terms of units, as `ms / (100/ms)` gives units of time squared, not Hz. The proper definition for your sampling frequency should be `int(1/dt)`.

5. Following up from point 4, when defining your simulation steps you write `sim_steps = int(T/dt) # number of simulation steps`. This evaluates to 150000, as you divide T (which you have initialized earlier to 15000, and given in ms) with dt (which is 0.1, and I assume is given in seconds). In reality, with a dt of 0.1s (100ms) you only have 10 steps per second, which would evaluate to `sim_steps = 150` (10 steps per second and 15 seconds of simulation time). Even if we assume a 10kHz sampling rate and ignore the above, if we follow the initialization of the sim_steps to 150000, then a single step does not correspond to a ms but 0.1ms, making all of the delays right below that (all the `td_...` variables) inconsistent. Overall, your comments and the declared variables create some confusion and could be improved.

6. Moreover, some of the figures that are produced by running the `DBS.py` file are strange. For example, in the figure `LFP - 80Hz.png` the x-axis I believe gives simulation steps instead of time. Furthermore, the figure with the LFP looks a bit strange (see the following figure). ![LFP - 80Hz](https://github.com/ReScience/submissions/assets/45063685/70a153f2-8219-43ec-a094-1a59bef53275). This figure was generated by running the `DBS.py` with the default parameters, and a DBS frequency of 150HZ (default value). Could you modify the code to be consistent with the rest of the plots?

7.  On another note, the simulations take an incredibly long time of around **2 hours** to complete on a Linux computer with a 10th gen 16-core Intel i7 processor with 64GB of RAM. Can you confirm that you are indeed running each simulation with the intended parameters? Also, if possible, could you specify the time it takes to run a complete simulation on your end, along with the technical specs of the system you used?

8. I noticed that in your results you are generating plots with the PSDs for DBS applied at 130Hz and 180Hz, however, the x-axis is clipped at 100Hz, thus I cannot verify the results. If you consider the results useful, please modify the x-axis so that we can see the results.

9. Finally, in the `model_functions.py` file, the function `PSD` that computes the periodogram using Welch's method (before plotting it using the `plot_PSD_DBS` function) and is being called during the `DBS.py` execution (lines 323, 328) also has hard-coded values for the `nperseg` parameter (line 222, `nperseg=10*1024`). Taken directly from the file:
```
def PSD(signal, fs):
    (f, S) = welch(signal, fs, nperseg=10*1024)

    return f, S
```
This should also be fixed as for the plotting functions.

# Minor language notes

I noticed that in your PDF, the spacing between paragraphs is not always consistent. Specifically, almost all paragraphs do not have vertical spacing, but the first paragraph in page 2 is different, as is the last paragraph of page 9. Is this done on purpose or is it an errand LaTeX error?

I am also listing some typos that I came across for your convenience. I would advise to proofread your text using a syntax / grammar checker (Grammarly, for example), to catch any others that I might have missed.

1. In section `3 Implementation details` - "...separete file" -> "separate file".

2. In section `3.3 The TCM model` - "...to devide and conquer..." -> "divide"

# Conclusion

I would like to thank the authors, as they took into account all earlier comments and made significant modifications to their work, implementing a range of suggestions and explaining their approach in a detailed and convincing manner. Importantly, they successfully replicated the claims of the original work regarding the effects of low-frequency DBS (<40 Hz) by using a stimulation protocol at 25Hz. Overall, the work represents a thorough replication of an interesting model to investigate DBS for PD.
